utils: (Worklist) move to utils
